+++ molgenis-semantic-search/src/main/java/org/molgenis/data/semanticsearch/config/SemanticSearchConfig.java
@@ -1,7 +1,7 @@
 package org.molgenis.data.semanticsearch.config;
 
 import org.molgenis.data.DataService;
-import org.molgenis.data.elasticsearch.config.ClientFactory;
+import org.molgenis.data.elasticsearch.client.ClientFactory;
 import org.molgenis.data.elasticsearch.generator.DocumentIdGenerator;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.TagFactory;
